requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.27.0
attrs==24.2.0
bcrypt==4.0.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email-validator==2.0.0
fastapi==0.103.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-anthropic==0.1.23
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.115
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openai==1.43.1
orjson==3.10.7
packaging==24.1
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
python-dotenv==1.0.1
python-ulid==1.1.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.15
starlette==0.27.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.15.0
yarl==1.9.11
